keyring prefix respect the new profiles
since we introduced profiles, we can have multiple config with the same hostname

to differentiate them apart, we append the profile name to the hostname that will be used to compute the hashcode value.

this value will be used as a integer prefix in the keyring.

This do break backwardscompability somewhat, however its a minor inconvience since it only mean the user need to re-authenticate one more time after the update sdpctl
after that, it will work as expected.